renamed request to query
